Update fast2.py
Fixed an error that only showed when the code was run for the first time. The error was generated when trying to cleanse a newly downloaded text file. I wanted the text file to be downloaded because it explains the data contained in the csv files should anybody be interested. I've set it now so that only csv files are cleansed.